Add Javadoc comments to generated Groovy interceptor classes
* Generalize some Javadoc utils
* Add Javadoc to the generated Groovy interceptor classes
* Minor refactoring of the Groovy generator
