+++ subprojects/internal-instrumentation-processor/src/main/java/org/gradle/internal/instrumentation/processor/codegen/groovy/InterceptGroovyCallsGenerator.java
@@ -25,7 +25,6 @@ import com.squareup.javapoet.TypeSpec;
 import org.gradle.internal.instrumentation.model.CallInterceptionRequest;
 import org.gradle.internal.instrumentation.model.CallableInfo;
 import org.gradle.internal.instrumentation.model.CallableKindInfo;
-import org.gradle.internal.instrumentation.model.ParameterKindInfo;
 import org.gradle.internal.instrumentation.model.RequestExtra;
 import org.gradle.internal.instrumentation.processor.codegen.InstrumentationCodeGenerator.GenerationResult.HasFailures.FailureInfo;
 import org.gradle.internal.instrumentation.processor.codegen.RequestGroupingInstrumentationClassGenerator;
@@ -39,19 +38,12 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Stack;
 import java.util.function.Consumer;
-import java.util.function.Function;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
-import static org.gradle.internal.instrumentation.processor.codegen.SignatureUtils.hasCallerClassName;
-import static org.gradle.internal.instrumentation.processor.codegen.groovy.ParameterMatchEntry.Kind.PARAMETER;
-import static org.gradle.internal.instrumentation.processor.codegen.groovy.ParameterMatchEntry.Kind.RECEIVER;
-import static org.gradle.internal.instrumentation.processor.codegen.groovy.ParameterMatchEntry.Kind.RECEIVER_AS_CLASS;
-import static org.gradle.internal.instrumentation.processor.codegen.groovy.ParameterMatchEntry.Kind.VARARG;
+import static org.gradle.internal.instrumentation.processor.codegen.JavadocUtils.callableKindForJavadoc;
+import static org.gradle.internal.instrumentation.processor.codegen.JavadocUtils.interceptedCallableLink;
+import static org.gradle.internal.instrumentation.processor.codegen.JavadocUtils.interceptorImplementationLink;
 
 public class InterceptGroovyCallsGenerator extends RequestGroupingInstrumentationClassGenerator {
     @Override
@@ -108,19 +100,13 @@ public class InterceptGroovyCallsGenerator extends RequestGroupingInstrumentatio
     }
 
     private static TypeSpec generateNamedCallableInterceptorClass(String name, List<CallInterceptionRequest> requests) {
-        SignatureTree signatureTree = signatureTreeFromRequests(requests);
-        return generateInterceptorClass(TextUtil.capitalize(name) + "CallInterceptor", namedCallableScopesArgs(name, requests), signatureTree).build();
+        String className = TextUtil.capitalize(name) + "CallInterceptor";
+        return generateInterceptorClass(className, namedCallableScopesArgs(name, requests), requests).build();
     }
 
     private static TypeSpec generateConstructorInterceptorClass(Type constructedType, List<CallInterceptionRequest> requests) {
-        SignatureTree signatureTree = signatureTreeFromRequests(requests);
-        TypeSpec.Builder generatedClass = generateInterceptorClass(
-            ClassName.bestGuess(constructedType.getClassName()).simpleName() + "ConstructorCallInterceptor",
-            constructorScopeArg(TypeUtils.typeName(constructedType)),
-            signatureTree
-        );
-
-        return generatedClass.build();
+        String className = ClassName.bestGuess(constructedType.getClassName()).simpleName() + "ConstructorCallInterceptor";
+        return generateInterceptorClass(className, constructorScopeArg(TypeUtils.typeName(constructedType)), requests).build();
     }
 
     private static SignatureTree signatureTreeFromRequests(Collection<CallInterceptionRequest> requests) {
@@ -129,9 +115,10 @@ public class InterceptGroovyCallsGenerator extends RequestGroupingInstrumentatio
         return result;
     }
 
-    private static TypeSpec.Builder generateInterceptorClass(String className, CodeBlock scopes, SignatureTree signatureTree) {
+    private static TypeSpec.Builder generateInterceptorClass(String className, CodeBlock scopes, List<CallInterceptionRequest> requests) {
         TypeSpec.Builder generatedClass = TypeSpec.classBuilder(className)
             .superclass(CALL_INTERCEPTOR_CLASS)
+            .addJavadoc(interceptorClassJavadoc(requests))
             .addModifiers(Modifier.PRIVATE, Modifier.STATIC);
 
         MethodSpec constructor = MethodSpec.constructorBuilder().addModifiers(Modifier.PUBLIC).addStatement("super($L)", scopes).build();
@@ -144,11 +131,21 @@ public class InterceptGroovyCallsGenerator extends RequestGroupingInstrumentatio
             .addParameter(INVOCATION_CLASS, "invocation")
             .addParameter(String.class, "consumer")
             .addException(Throwable.class)
-            .addCode(generateCodeFromInterceptorSignatureTree(signatureTree))
+            .addCode(generateCodeFromInterceptorSignatureTree(signatureTreeFromRequests(requests)))
             .build();
 
         generatedClass.addMethod(doIntercept);
         return generatedClass;
+    }
+
+    private static CodeBlock interceptorClassJavadoc(Collection<CallInterceptionRequest> requests) {
+        List<CodeBlock> result = new ArrayList<>();
+        result.add(CodeBlock.of("Intercepts the following declarations:<ul>"));
+        requests.stream().map(request ->
+            CodeBlock.of("<li> $L $L\n     with $L", callableKindForJavadoc(request), interceptedCallableLink(request), interceptorImplementationLink(request))
+        ).collect(Collectors.toCollection(() -> result));
+        result.add(CodeBlock.of("</ul>"));
+        return result.stream().collect(CodeBlock.joining("\n\n"));
     }
 
     private static CodeBlock constructorScopeArg(TypeName constructedType) {
@@ -177,131 +174,7 @@ public class InterceptGroovyCallsGenerator extends RequestGroupingInstrumentatio
         CodeBlock.Builder result = CodeBlock.builder();
         result.addStatement("$T receiver = invocation.getReceiver()", Object.class);
 
-        class CodeGeneratingSignatureTreeVisitor {
-            private final Stack<CodeBlock> paramVariablesStack = new Stack<>();
-
-            /**
-             * @param paramIndex index of the parameter in the signatures, -1 stands for the receiver
-             */
-            void visit(SignatureTree current, int paramIndex) {
-                CallInterceptionRequest leafInCurrent = current.getLeafOrNull();
-                if (leafInCurrent != null) {
-                    generateInvocationWhenArgsMatched(leafInCurrent, paramIndex);
-                }
-                Map<ParameterMatchEntry, SignatureTree> children = current.getChildrenByMatchEntry();
-                if (!children.isEmpty()) {
-                    boolean hasParamMatchers = children.keySet().stream().anyMatch(it -> it.kind == PARAMETER);
-                    if (hasParamMatchers) { // is not the receiver or vararg
-                        result.beginControlFlow("if (invocation.getArgsCount() > $L)", paramIndex);
-                        result.addStatement("Object arg$1L = invocation.getArgument($1L)", paramIndex);
-                    }
-                    // Visit non-vararg invocations first and varargs after:
-                    children.forEach((entry, child) -> {
-                        if (entry.kind != VARARG) {
-                            generateNormalCallChecksAndVisitSubtree(entry, child, paramIndex);
-                        }
-                    });
-                    if (hasParamMatchers) {
-                        result.endControlFlow();
-                    }
-                    children.forEach((entry, child) -> {
-                        if (entry.kind == VARARG) {
-                            generateVarargCheckAndInvocation(entry, child, paramIndex);
-                        }
-                    });
-                }
-            }
-
-            private void generateInvocationWhenArgsMatched(CallInterceptionRequest request, int argCount) {
-                TypeName implementationOwner = TypeUtils.typeName(request.getImplementationInfo().getOwner());
-                result.beginControlFlow("if (invocation.getArgsCount() == $L)", argCount);
-                CodeBlock argsCode = prepareInvocationArgs(request);
-                emitInvocationCodeWithReturn(request, implementationOwner, argsCode);
-                result.endControlFlow();
-            }
-
-            private CodeBlock prepareInvocationArgs(CallInterceptionRequest request) {
-                boolean hasKotlinDefaultMask = request.getInterceptedCallable().getParameters().stream().anyMatch(it -> it.getKind() == ParameterKindInfo.KOTLIN_DEFAULT_MASK);
-                boolean hasCallerClassName = hasCallerClassName(request.getInterceptedCallable());
-                Stream<CodeBlock> maybeZeroForKotlinDefault = hasKotlinDefaultMask ? Stream.of(CodeBlock.of("0")) : Stream.empty();
-                Stream<CodeBlock> maybeCallerClassName = hasCallerClassName ? Stream.of(CodeBlock.of("consumer")) : Stream.empty();
-                return Stream.of(
-                    paramVariablesStack.stream(),
-                    maybeZeroForKotlinDefault,
-                    maybeCallerClassName
-                ).flatMap(Function.identity()).collect(CodeBlock.joining(", "));
-            }
-
-            private void emitInvocationCodeWithReturn(CallInterceptionRequest request, TypeName implementationOwner, CodeBlock argsCode) {
-                String implementationName = request.getImplementationInfo().getName();
-                if (request.getInterceptedCallable().getKind() == CallableKindInfo.AFTER_CONSTRUCTOR) {
-                    result.addStatement("$1T result = new $1T($2L)", TypeUtils.typeName(request.getInterceptedCallable().getOwner()), paramVariablesStack.stream().collect(CodeBlock.joining(", ")));
-                    CodeBlock interceptorArgs = CodeBlock.join(Arrays.asList(CodeBlock.of("result"), argsCode), ", ");
-                    result.addStatement("$T.$L($L)", implementationOwner, implementationName, interceptorArgs);
-                    result.addStatement("return result");
-                } else if (request.getInterceptedCallable().getReturnType() == Type.VOID_TYPE) {
-                    result.addStatement("$T.$L($L)", implementationOwner, implementationName, argsCode);
-                    result.addStatement("return null");
-                } else {
-                    result.addStatement("return $T.$L($L)", implementationOwner, implementationName, argsCode);
-                }
-            }
-
-            private void generateVarargCheckAndInvocation(ParameterMatchEntry entry, SignatureTree child, int paramIndex) {
-                TypeName entryParamType = TypeUtils.typeName(entry.type);
-
-                result.add("// Trying to match the vararg invocation\n");
-                CodeBlock varargVariable = CodeBlock.of("varargValues");
-                result.addStatement("$1T[] $2L = new $1T[invocation.getArgsCount() - $3L]", entryParamType, varargVariable, paramIndex);
-                CodeBlock varargMatched = CodeBlock.of("varargMatched");
-                result.addStatement("boolean $L = true", varargMatched);
-                result.beginControlFlow("for (int argIndex = $1L; argIndex < invocation.getArgsCount(); argIndex++)", paramIndex);
-
-                CodeBlock nextArg = CodeBlock.of("nextArg");
-                result.addStatement("Object $L = invocation.getArgument(argIndex)", nextArg);
-                result.beginControlFlow("if ($L instanceof $T)", nextArg, entryParamType);
-                result.addStatement("$1L[argIndex - $2L] = ($3T) $4L", varargVariable, paramIndex, entryParamType, nextArg);
-                result.nextControlFlow("else");
-                result.addStatement("$L = false", varargMatched);
-                result.addStatement("break");
-                result.endControlFlow();
-
-                result.endControlFlow();
-                result.beginControlFlow("if ($L)", varargMatched);
-                paramVariablesStack.push(varargVariable);
-                CallInterceptionRequest request = Objects.requireNonNull(child.getLeafOrNull());
-                emitInvocationCodeWithReturn(request, TypeUtils.typeName(request.getImplementationInfo().getOwner()), prepareInvocationArgs(request));
-                paramVariablesStack.pop();
-                result.endControlFlow();
-            }
-
-            private void generateNormalCallChecksAndVisitSubtree(ParameterMatchEntry entry, SignatureTree child, int paramIndex) {
-                CodeBlock argExpr = entry.kind == RECEIVER || entry.kind == RECEIVER_AS_CLASS
-                    ? CodeBlock.of("receiver")
-                    : CodeBlock.of("arg$L", paramIndex);
-
-                int childArgCount = paramIndex + 1;
-                TypeName entryChildType = TypeUtils.typeName(entry.type);
-                CodeBlock matchExpr = entry.kind == RECEIVER_AS_CLASS ?
-                    CodeBlock.of("$L.equals($T.class)", argExpr, entryChildType) :
-                    // Vararg fits here, too:
-                    CodeBlock.of("$L instanceof $T", argExpr, entryChildType.box());
-                result.beginControlFlow("if ($L)", matchExpr);
-                boolean shouldPopParameter = false;
-                if (entry.kind != RECEIVER_AS_CLASS) {
-                    shouldPopParameter = true;
-                    CodeBlock paramVariable = CodeBlock.of("$LTyped", argExpr);
-                    result.addStatement("$2T $1L = ($2T) $3L", paramVariable, entryChildType, argExpr);
-                    paramVariablesStack.push(paramVariable);
-                }
-                visit(child, childArgCount);
-                if (shouldPopParameter) {
-                    paramVariablesStack.pop();
-                }
-                result.endControlFlow();
-            }
-        }
-        new CodeGeneratingSignatureTreeVisitor().visit(tree, -1);
+        new CodeGeneratingSignatureTreeVisitor(result).visit(tree, -1);
 
         result.addStatement("return invocation.callOriginal()");
         return result.build();
